Updates to esbuild
Adds esbuild as a development dependency.

Adds a esbuild custom build script allowing us to create an esbuild
plugin.

This plugin defines a custom require/import rule for any pattern that
matches `name-prebuilt` such that when that name is required we search
for any prebuilt libraries and "bundle" them into the code. These are
included as base64 encoded strings.

When the code is run via node, these strings are iterated, written to a
temp file, and then are attempted to be required from this temporary
file. If the native library, now on disk, can be loaded we use that as
the required library.

Since node's `require` statement uses a cache this ensures that the
native library is already loaded by the system PRIOR to the dependent
library attempting to load it. This allows us to effectively bundle the
native libraries.

For example the `usb` package contains native prebuilt libraries. If we
were to simply `require("usb")` it would fail when our bundled code runs
as we no longer have the prebuilt files sitting in `node_modules`

However, if we first `reuiqre("usb-prebuilt")` then the esbuild process
will find those libraries at build-time and include them. Naturally this
makes our bundle quite large. However, when the code runs we will be
able to find, load, and cache the necessary native library. The code the
user would use is:

// Load the libraries
require("usb-prebuilt")

// Load the actual package
const usb = require("usb")